implement: story-feature-split-cli-daemon-commands-module
