+++ crates/cli/src/entrypoint.rs
@@ -32,6 +32,9 @@ use crate::{
     dispatch::Dispatch,
 };
 
+#[path = "daemon_commands.rs"]
+mod daemon_commands;
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum OutputChannel {
     Stdout,
@@ -257,9 +260,10 @@ where
         Dispatch::HarnessDisable(args) => {
             (execute_system_harness_disable(&args), OutputChannel::Stdout)
         }
-        Dispatch::DaemonEnable(args) => {
-            (execute_system_daemon_enable(&args), OutputChannel::Stdout)
-        }
+        Dispatch::DaemonEnable(args) => (
+            daemon_commands::execute_system_daemon_enable(&args),
+            OutputChannel::Stdout,
+        ),
         Dispatch::DaemonDisable(args) => {
             (execute_system_daemon_disable(&args), OutputChannel::Stdout)
         }
@@ -283,171 +287,6 @@ fn execute_system_daemon_run(_args: &crate::command::OutputArgs) -> Outcome {
     execute_system_reconciliation("daemon run", |application| {
         application.execute_daemon_cycle()
     })
-}
-
-fn execute_system_daemon_enable(args: &crate::command::DaemonEnableArgs) -> Outcome {
-    let command = "daemon enable";
-    let paths = match PlatformPaths::resolve(&ProcessEnvironment) {
-        Ok(paths) => paths,
-        Err(_) => return repository_composition_error(command),
-    };
-    let filesystem = SystemFileSystem;
-    let config = match FileConfigRepository::new(&filesystem, paths.skilltap_config().clone())
-        .and_then(|repository| match repository.load() {
-            Ok(DocumentState::Present(config)) => Ok(config),
-            Ok(DocumentState::Missing) => Ok(ConfigDocument::defaults()),
-            Err(error) => Err(error),
-        }) {
-        Ok(config) => config,
-        Err(_) => return repository_composition_error(command),
-    };
-    let interval = args.interval.unwrap_or(config.updates().interval);
-    let platform = crate::daemon::platform(&paths);
-    let executable = match std::env::current_exe()
-        .ok()
-        .and_then(|path| path.to_str().map(str::to_owned))
-        .and_then(|path| AbsolutePath::new(path).ok())
-    {
-        Some(path) => path,
-        None => {
-            return Outcome::new(command, ResultClass::Invalid).with_error(ErrorDetail::new(
-                "daemon_executable_unavailable",
-                "The skilltap executable path could not be represented safely.",
-            ));
-        }
-    };
-    let definition =
-        match skilltap_core::daemon::render_service(&skilltap_core::daemon::DaemonServiceSpec {
-            platform,
-            interval,
-            executable,
-        }) {
-            Ok(definition) => definition,
-            Err(_) => {
-                return Outcome::new(command, ResultClass::Invalid).with_error(ErrorDetail::new(
-                    "daemon_definition_invalid",
-                    "The daemon service definition could not be validated.",
-                ));
-            }
-        };
-    let files = crate::daemon::files(&paths, &definition);
-    let root = crate::daemon::root(&paths, platform);
-    if let Err(_error) = filesystem.create_directory_all(&root) {
-        return Outcome::new(command, ResultClass::AttentionRequired).with_warning(
-            Warning::new(
-                "daemon_definition_write_failed",
-                "The daemon service directory could not be created.",
-            )
-            .with_context("path", root.as_str()),
-        );
-    }
-    let mut existing = Vec::with_capacity(files.len());
-    for (path, _file) in &files {
-        match filesystem.read_regular_no_follow(path) {
-            Ok(Some(contents))
-                if crate::daemon::owns(platform, &contents)
-                    && !crate::daemon::valid(platform, &contents) =>
-            {
-                return Outcome::new(command, ResultClass::AttentionRequired)
-                    .with_resource(OutputEntry::new(path.as_str(), "malformed"))
-                    .with_warning(Warning::new(
-                        "daemon_definition_malformed",
-                        "An owned daemon service definition is malformed; it was not replaced.",
-                    ));
-            }
-            Ok(Some(contents)) if !crate::daemon::owns(platform, &contents) => {
-                return Outcome::new(command, ResultClass::AttentionRequired)
-                    .with_resource(OutputEntry::new(path.as_str(), "conflict"))
-                    .with_warning(Warning::new(
-                        "daemon_definition_conflict",
-                        "An unmanaged service definition already occupies the skilltap path.",
-                    ));
-            }
-            Ok(value) => existing.push(value),
-            Err(_) => {
-                return Outcome::new(command, ResultClass::AttentionRequired).with_warning(
-                    Warning::new(
-                        "daemon_definition_unreadable",
-                        "The existing daemon service definition could not be read safely.",
-                    )
-                    .with_context("path", path.as_str()),
-                );
-            }
-        }
-    }
-    let changed_files = files
-        .iter()
-        .zip(existing.iter())
-        .filter(|((_, file), current)| current.as_deref() != Some(file.contents().as_bytes()))
-        .map(|((path, file), current)| {
-            (
-                path.clone(),
-                file.contents().as_bytes().to_vec(),
-                current.clone(),
-            )
-        })
-        .collect::<Vec<_>>();
-    if let Err((path, _error)) = publish_daemon_files(&filesystem, &changed_files) {
-        return Outcome::new(command, ResultClass::AttentionRequired)
-            .with_summary("changed", false)
-            .with_warning(
-                Warning::new(
-                    "daemon_definition_write_failed",
-                    "The daemon service definition could not be published atomically; prior files were restored.",
-                )
-                .with_context("path", path.as_str()),
-            );
-    }
-    if run_service_manager(platform, ServiceManagerAction::Enable, &files[0].0).is_err() {
-        return Outcome::new(command, ResultClass::AttentionRequired)
-            .with_summary("changed", !changed_files.is_empty())
-            .with_warning(Warning::new(
-                "daemon_manager_unavailable",
-                "The service definition was written, but the user service manager did not activate it.",
-            ))
-            .with_next_action(NextAction::new(
-                "retry_daemon_enable",
-                "Retry daemon enable after checking the user service manager.",
-            ))
-            .with_resource(OutputEntry::new(files[0].0.as_str(), "installed"));
-    }
-    Outcome::new(command, ResultClass::Completed)
-        .with_resource(
-            OutputEntry::new(files[0].0.as_str(), "enabled")
-                .with_field("interval", interval.to_string())
-                .with_field("platform", format!("{platform:?}").to_lowercase()),
-        )
-        .with_summary("changed", !changed_files.is_empty())
-}
-
-type DaemonChangedFile = (AbsolutePath, Vec<u8>, Option<Vec<u8>>);
-
-/// Publish all changed service definitions as one recoverable pair. If a
-/// later definition fails, every earlier write is restored to its prior bytes
-/// (or removed when it did not previously exist).
-fn publish_daemon_files(
-    filesystem: &dyn FileSystem,
-    changed_files: &[DaemonChangedFile],
-) -> Result<(), (AbsolutePath, skilltap_core::runtime::RuntimeError)> {
-    let mut written: Vec<&DaemonChangedFile> = Vec::new();
-    for changed in changed_files {
-        let (path, contents, _previous) = changed;
-        if let Err(error) = filesystem.atomic_write(path, contents) {
-            for (written_path, _, written_previous) in written.iter().rev().copied() {
-                match written_previous {
-                    Some(previous) => {
-                        let _ = filesystem.atomic_write(written_path, previous);
-                    }
-                    None => {
-                        let _ = filesystem.remove(written_path);
-                    }
-                }
-            }
-            return Err((path.clone(), error));
-        }
-        written.push(changed);
-    }
-    Ok(())
 }
 
 fn execute_system_daemon_disable(_args: &OutputArgs) -> Outcome {
