+++ crates/cli/src/entrypoint/tests.rs
@@ -1,4 +1,4 @@
-use std::{cell::RefCell, collections::BTreeMap, ffi::OsString, io};
+use std::ffi::OsString;
 
 use serde_json::Value;
 
@@ -169,122 +169,4 @@ fn help_and_version_complete_on_stdout() {
         assert_eq!(execution.channel, OutputChannel::Stdout);
         assert!(!execution.document.is_empty());
     }
-}
-
-struct FailingPublicationFileSystem {
-    files: RefCell<BTreeMap<String, Vec<u8>>>,
-    fail_path: String,
-}
-
-impl FailingPublicationFileSystem {
-    fn new(fail_path: &str) -> Self {
-        Self {
-            files: RefCell::new(BTreeMap::new()),
-            fail_path: fail_path.to_owned(),
-        }
-    }
-}
-
-impl skilltap_core::runtime::FileSystem for FailingPublicationFileSystem {
-    fn inspect(
-        &self,
-        _path: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<skilltap_core::runtime::FileMetadata, skilltap_core::runtime::RuntimeError> {
-        unimplemented!("publication helper does not inspect")
-    }
-
-    fn canonicalize(
-        &self,
-        _path: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<skilltap_core::domain::AbsolutePath, skilltap_core::runtime::RuntimeError> {
-        unimplemented!("publication helper does not canonicalize")
-    }
-
-    fn create_directory_all(
-        &self,
-        _path: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<(), skilltap_core::runtime::RuntimeError> {
-        unimplemented!("publication helper does not create directories")
-    }
-
-    fn read(
-        &self,
-        _path: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<Vec<u8>, skilltap_core::runtime::RuntimeError> {
-        unimplemented!("publication helper does not read")
-    }
-
-    fn read_regular_no_follow(
-        &self,
-        _path: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<Option<Vec<u8>>, skilltap_core::runtime::RuntimeError> {
-        unimplemented!("publication helper does not inspect regular files")
-    }
-
-    fn atomic_write(
-        &self,
-        path: &skilltap_core::domain::AbsolutePath,
-        contents: &[u8],
-    ) -> Result<(), skilltap_core::runtime::RuntimeError> {
-        if path.as_str() == self.fail_path {
-            return Err(skilltap_core::runtime::RuntimeError::FileSystem {
-                action: skilltap_core::runtime::FileSystemAction::Write,
-                path: path.clone(),
-                source: io::Error::other("injected second-write failure"),
-            });
-        }
-        self.files
-            .borrow_mut()
-            .insert(path.as_str().to_owned(), contents.to_vec());
-        Ok(())
-    }
-
-    fn copy_recoverable(
-        &self,
-        _source: &skilltap_core::domain::AbsolutePath,
-        _destination: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<(), skilltap_core::runtime::RuntimeError> {
-        unimplemented!("publication helper does not copy")
-    }
-
-    fn create_relative_symlink(
-        &self,
-        _target: &skilltap_core::runtime::RelativeSymlinkTarget,
-        _link: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<(), skilltap_core::runtime::RuntimeError> {
-        unimplemented!("publication helper does not symlink")
-    }
-
-    fn remove(
-        &self,
-        path: &skilltap_core::domain::AbsolutePath,
-    ) -> Result<(), skilltap_core::runtime::RuntimeError> {
-        self.files.borrow_mut().remove(path.as_str());
-        Ok(())
-    }
-}
-
-#[test]
-fn daemon_pair_publication_restores_earlier_service_files_on_later_failure() {
-    let service = skilltap_core::domain::AbsolutePath::new("/tmp/skilltap-service").unwrap();
-    let timer = skilltap_core::domain::AbsolutePath::new("/tmp/skilltap-timer").unwrap();
-    let filesystem = FailingPublicationFileSystem::new(timer.as_str());
-    filesystem
-        .files
-        .borrow_mut()
-        .insert(service.as_str().to_owned(), b"old service".to_vec());
-
-    let changed = vec![
-        (
-            service.clone(),
-            b"new service".to_vec(),
-            Some(b"old service".to_vec()),
-        ),
-        (timer.clone(), b"new timer".to_vec(), None),
-    ];
-    let error = publish_daemon_files(&filesystem, &changed).unwrap_err();
-    assert_eq!(error.0, timer);
-    let files = filesystem.files.borrow();
-    assert_eq!(files.get(service.as_str()), Some(&b"old service".to_vec()));
-    assert!(!files.contains_key(timer.as_str()));
 }
